requirements.txt
Flask
PyMySQL
Werkzeug
PyJWT
